chore: start module renaming
rex.py --> rex_wrapper.py
explanation.py --> rex.py
