+++ tests/snapshot_tests/_explanation_test.py
@@ -1,5 +1,5 @@
 import pytest
-from rex_xai.explanation import _explanation, analyze
+from rex_xai.rex import _explanation, analyze
 from rex_xai._utils import Strategy
 
 @pytest.mark.parametrize("batch_size", [1, 64])
