+++ tests/unit_tests/multi-explanation_test.py
@@ -4,7 +4,7 @@ import os
 import numpy as np
 import pytest
 from rex_xai._utils import Strategy
-from rex_xai.explanation import calculate_responsibility
+from rex_xai.rex import calculate_responsibility
 from rex_xai.extraction import Explanation
 from rex_xai.multi_explanation import MultiExplanation
